requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyarrow>=14.0
requests>=2.31
pyyaml>=6.0
joblib>=1.3
streamlit>=1.30
plotly>=5.18
sportsdataverse>=0.0.36
xgboost>=2.0,<3.0
matplotlib>=3.7
setuptools<70
